Added Flow Control Projects directory with initial projects and README file
This commit adds the flow-control_projects directory to the repository, which contains a collection of programming projects that focus on flow control statements in the C programming language. The README file provides an overview of the directory's contents, including a list of projects and instructions for getting started and contributing. This commit also includes the initial set of projects, including Even or Odd, Factorial, Fibonacci Series, Prime Number Checker, and Reverse Number